\name{addhull}
\alias{addhull}
\title{Add convex hulls}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add conver hulls using the tool qconvex
}
\usage{addhull(gd, g, by)}
\arguments{
\item{gd}{ggobi dataset}
\item{g}{ggobi reference}
\item{by}{grouping variable}
}

\details{To use this command you must have qconvex installed and available
on your path.   I'm not sure if this will work on windows (probably not)
but it's not a big loss, because the technique isn't very useful
anyway.}

\examples{}
\keyword{hplot}
