\name{as.data.frame.clusterfly}
\alias{as.data.frame.clusterfly}
\title{Convert clusterfly object to data.frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Concatenates data and cluster assignments into one data.frame.  Cluster assignments are prefixed with \code{cl_}
}
\usage{as.data.frame.clusterfly(x, ...)}
\arguments{
\item{x}{clusterfly object}
\item{...}{}
}

\details{}

\examples{}
\keyword{manip}
