\name{as.data.frame.clusterfly}
\alias{as.data.frame.clusterfly}
\title{Convert clusterfly object to data...}
\usage{\method{as.data.frame}{clusterfly} (x, ...)}

\description{
  Convert clusterfly object to data.frame. Concatenates data and
  cluster assignments into one data.frame.  Cluster assignments are
  prefixed with \code{cl_}.
}
\keyword{manip}
\arguments{
  \item{x}{clusterfly object}
  \item{...}{ignored}
}
