\name{cfly_show}
\alias{cfly_show}
\title{Dynamic plot: show in ggobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Opens an instance ggobi for this dataset (if not already open), and colours the points according the cluster assignment.
}
\usage{cfly_show(cf, idx = "true", hulls = FALSE)}
\arguments{
\item{cf}{clusterfly object}
\item{idx}{clustering to display}
\item{hulls}{add convex hull? see \code{\link{addhull}} for details}
}

\details{}

\examples{}
\keyword{dynamic}
