\name{clarify}
\alias{clarify}
\title{Clarify matrix}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Clarify matrix ordering to minimize off diagonals
}
\usage{clarify(a, b, method="greedy")}
\arguments{
\item{a}{cluster assignments to reassign}
\item{b}{matrix b}
\item{method}{}
}
\value{vector of reassigned cluster a}
\details{}
\seealso{\code{\link[e1071]{matchClasses}}}
\examples{}
\keyword{manip}
