% Generated by roxygen2 (4.0.0): do not edit by hand
\name{clarify}
\alias{clarify}
\title{Clarify matrix
Clarify matrix ordering to minimize off diagonals}
\usage{
clarify(a, b, method = "greedy")
}
\arguments{
\item{a}{cluster assignments to reassign}

\item{b}{matrix b}

\item{method}{clarification method}
}
\value{
vector of reassigned cluster a
}
\description{
Clarify matrix
Clarify matrix ordering to minimize off diagonals
}
\seealso{
\code{\link[e1071]{matchClasses}}
}
\keyword{manip}

