\name{clusters}
\alias{clusters}
\alias{clusters.kmeans}
\alias{clusters.default}
\title{Get clusters}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Extract clusters from clustering object
}
\usage{clusters(x)}
\arguments{
\item{x}{object}
}

\details{}

\examples{}
\keyword{internal}
