\name{combinerows}
\alias{combinerows}
\title{Combine multiple rows in cluster hierarchy to make intermediate node}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
@keyword internal
}
\usage{combinerows(df, cat)}
\arguments{
\item{df}{}
\item{cat}{}
}

\details{}

\examples{}
\keyword{internal}
