\name{hierarchical}
\alias{hierarchical}
\title{Hierachical clustering}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenient methods for hierachical clustering
}
\usage{hierarchical(df, method="complete", metric="euclidean", n=5)}
\arguments{
\item{df}{data frame}
\item{method}{method to use, see \code{\link{hclust}}}
\item{metric}{distance metric to use, see \code{\link{dist}}}
\item{n}{number of clusters to retrieve, see \code{\link{cut}}}
}

\details{}

\examples{}
\keyword{cluster}
