% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addhull}
\alias{addhull}
\title{Add convex hulls
Add conver hulls using the tool qconvex}
\usage{
addhull(gd, g, by)
}
\arguments{
\item{gd}{ggobi dataset}

\item{g}{ggobi reference}

\item{by}{grouping variable}
}
\description{
To use this command you must have qconvex installed and available
on your path.   I'm not sure if this will work on windows (probably not)
but it's not a big loss, because the technique isn't very useful
anyway.
}
\keyword{hplot}

