\name{addhull}
\alias{addhull}
\title{Add convex hulls...}
\usage{addhull(gd, g, by)}

\description{
  Add convex hulls Add conver hulls using the tool qconvex
}

\details{
  To use this command you must have qconvex installed and available on
  your path.  I'm not sure if this will work on windows (probably not)
  but it's not a big loss, because the technique isn't very useful
  anyway.
}
\keyword{hplot}
\arguments{
  \item{gd}{ggobi dataset}
  \item{g}{ggobi reference}
  \item{by}{grouping variable}
}
