% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.data.frame.clusterfly}
\alias{as.data.frame.clusterfly}
\title{Convert clusterfly object to data.frame.
Concatenates data and cluster assignments into one data.frame.
Cluster assignments are prefixed with \code{cl_}.}
\usage{
\method{as.data.frame}{clusterfly}(x, ...)
}
\arguments{
\item{x}{clusterfly object}

\item{...}{ignored}
}
\description{
Convert clusterfly object to data.frame.
Concatenates data and cluster assignments into one data.frame.
Cluster assignments are prefixed with \code{cl_}.
}
\keyword{manip}

