% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cfly_animate}
\alias{cfly_animate}
\title{Dynamic plot: Animate glyph colours}
\usage{
cfly_animate(cf, clusters = seq_along(cf$clusters), pause = 1,
  print = TRUE, max_iterations = 100)
}
\arguments{
\item{cf}{list of cluster ids that you want to animate across}

\item{clusters}{clusters to display}

\item{pause}{clusters number of seconds to pause between each change}

\item{print}{print current cluster to screen?}

\item{max_iterations}{maximum number of interations}
}
\description{
This function will animate until you manually break the loop
using Ctrl-Break or Ctrl-C.
}
\examples{
# Press Ctrl-Break or Ctrl-C to exit
if (interactive()) {
o <- olive_example()
cfly_animate(cfly_clarify(o), max = 5)
close(o)
}
}
\keyword{dynamic}

