\name{clarify}
\alias{clarify}
\title{Clarify matrix...}
\usage{clarify(a, b, method="greedy")}

\description{
  Clarify matrix Clarify matrix ordering to minimize off diagonals
}
\value{vector of reassigned cluster a}
\keyword{manip}
\seealso{\code{\link[e1071]{matchClasses}}}
\arguments{
  \item{a}{cluster assignments to reassign}
  \item{b}{matrix b}
  \item{method}{clarification method}
}
