\name{hierarchical}
\alias{hierarchical}
\title{Hierachical clustering...}
\usage{hierarchical(df, method="complete", metric="euclidean", n=5)}

\description{
  Hierachical clustering Convenient methods for hierachical clustering
}
\keyword{cluster}
\arguments{
  \item{df}{data frame}
  \item{method}{method to use, see \code{\link{hclust}}}
  \item{metric}{distance metric to use, see \code{\link{dist}}}
  \item{n}{number of clusters to retrieve, see \code{\link{cut}}}
}
