% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hierarchical}
\alias{hierarchical}
\title{Hierachical clustering
Convenient methods for hierachical clustering}
\usage{
hierarchical(df, method = "complete", metric = "euclidean", n = 5)
}
\arguments{
\item{df}{data frame}

\item{method}{method to use, see \code{\link{hclust}}}

\item{metric}{distance metric to use, see \code{\link{dist}}}

\item{n}{number of clusters to retrieve, see \code{\link{cut}}}
}
\description{
Hierachical clustering
Convenient methods for hierachical clustering
}
\keyword{cluster}

